Acquia Drupal 7.0.0 (Drupal 7.0 core) has just been released.
* Please reference the Getting Started Guide  
(http://network.acquia.com/documentation/getting-started/introduction) 
for more information.

Acquia Drupal consists of Acquia's package of Drupal core with a curated set  
of contributed modules, plus the ability to connect to the Acquia Network.

Notes:


* Acquia Drupal may include released, beta, alpha, or patched modules  
depending on which version is determined to be the most current, stable,  
secure and appropriate version for general use.

* Any patches to drupal core or individual modules are listed in a  
PATCHES.txt file with each one.



git-svn-id: file:///mnt/www/svn/drupal/branches/7.x@5427 08d20552-8ba5-4116-8c9e-e96c1c1f3480